+++ src/fcoverage/tasks/__init__.py
@@ -1,10 +1,10 @@
 # This file makes the 'tasks' directory a Python package.
 from .feature_extraction import FeatureExtractionTask
 from .analysis_tests import AnalyseTestsTask
-from .sourcecode_embedding import SourceCodeEmbeddingTask
+from .code_analysis import CodeAnalysisTask
 
 __all__ = [
     "FeatureExtractionTask",
     "AnalyseTestsTask",
-    "SourceCodeEmbeddingTask",
+    "CodeAnalysisTask",
 ]
